Implement shopping cart functionality with AJAX

+++ coffee-shop/database/migrations/2023_11_20_101907_create_products_table.php
@@ -15,7 +15,7 @@ return new class extends Migration
             $table->id();
             $table->string('name');
             $table->longText('description')->nullable();
-            $table->integer('category_id');
+            $table->bigInteger('category_id');
             $table->string('image');
             $table->integer('price')->default(0);
             $table->string('status')->default('pending');
